fix run command
